requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
supabase
vecs

streamlit
streamlit-option-menu
streamlit-authenticator
streamlit-extras